`#330 Introduced the concept of factory for CqlEngine and moved related methods to it
